+++ src/app/features/admin/danhmuc/coquanquanly/coquanquanly-list/coquanquanly-list.component.ts
@@ -142,6 +142,9 @@ export class DmCoquanquanlyListComponent implements OnInit {
   * Hàm lấy dữ liệu Cơ Quan Quản Lý
   */
   async getAllCoQuanQuanLy(param: any = { PageNumber: 1, PageSize: -1 }) {
+    if (this.listCoQuanQuanLy != null && this.listCoQuanQuanLy.length > 0) {
+      this.gridCoQuanQuanLy.clearSelection();
+    }
     const listData: any = await this.dmFacadeService
       .getDmCoQuanQuanLyService()
       .getFetchAll(param);
@@ -157,17 +160,14 @@ export class DmCoquanquanlyListComponent implements OnInit {
    * Hàm load lại dữ liệu grid
    */
   public reloadDataGrid() {
-  if (this.listCoQuanQuanLy.length > 0) {
-    this.gridCoQuanQuanLy.clearSelection();
-  }
-  this.formSearch.reset({
-    Keyword: "",
-    Idtinh: "",
-    Idhuyen: "",
-    Idxa: "",
-    Trangthai: ""
-  });
-  this.getAllCoQuanQuanLy();
+    this.formSearch.reset({
+      Keyword: "",
+      Idtinh: "",
+      Idhuyen: "",
+      Idxa: "",
+      Trangthai: ""
+    });
+    this.getAllCoQuanQuanLy();
   }
 
   /**
@@ -198,9 +198,6 @@ export class DmCoquanquanlyListComponent implements OnInit {
    * Tìm kiếm nâng cao
    */
   public searchAdvance() {
-    if (this.listCoQuanQuanLy.length > 0) {
-      this.gridCoQuanQuanLy.clearSelection();
-    }
     let dataSearch = this.formSearch.value;
     dataSearch['PageNumber'] = Paging.PageNumber;
     dataSearch['PageSize'] = Paging.PageSize;
@@ -263,6 +260,7 @@ export class DmCoquanquanlyListComponent implements OnInit {
    * Hàm delete mảng item đã chọn
    */
   public deleteArrayItem() {
+    let idItems: string[] = [];
     const dialogRef = this.commonService.confirmDeleteDiaLogService("", this.dataTranslate.DANHMUC.coquanquanly.confirmedContentOfDeleteDialog);
     dialogRef.afterClosed().subscribe(async (result) => {
       if (result === "confirm") {
@@ -274,8 +272,29 @@ export class DmCoquanquanlyListComponent implements OnInit {
             this.dataTranslate.DANHMUC.coquanquanly.nameofobject + " (" + data.tencoquanquanly + ") " + this.dataTranslate.DANHMUC.coquanquanly.informedContentOfUnDeletedDialog,
             this.dataTranslate.DANHMUC.coquanquanly.informedDialogTitle,
           );
+        } else {
+          this.listDataSelect.map(res => {
+            idItems.push(res.idcoquanquanly);
+          });
 
-          informationDialogRef.afterClosed().subscribe(() => {});
+          const dataBody: any = {
+            listId: idItems,
+          };
+
+          this.dmFacadeService.getDmCoQuanQuanLyService()
+          .deleteItemsCoQuanQuanLy(dataBody)
+          .subscribe(
+            () => {
+              this.getAllCoQuanQuanLy();
+            },
+            (error: HttpErrorResponse) => {
+              this.commonService.showeNotiResult(error.message, 2000);
+            },
+            () =>
+              this.commonService.showeNotiResult(
+                this.dataTranslate.COMMON.default.successDelete,
+                2000
+            ));
         }
       }
     });
@@ -402,20 +421,30 @@ export class DmCoquanquanlyListComponent implements OnInit {
     );
     dialogRef.afterClosed().subscribe(async (result) => {
       if (result === "confirm") {
-        await this.dmFacadeService
-          .getDmCoQuanQuanLyService()
-          .deleteItem({ idCoquanquanly: this.selectedItem.idcoquanquanly })
-          .subscribe(
-            () => this.getAllCoQuanQuanLy(),
-            (error: HttpErrorResponse) => {
-              this.commonService.showeNotiResult(error.message, 2000);
-            },
-            () =>
-              this.commonService.showeNotiResult(
-                this.dataTranslate.COMMON.default.successDelete,
-                2000
-              )
+        const data = this.generalClientService.findByKeyName<any>([this.selectedItem], "trangthai", TrangThaiEnum.Active);
+
+        if (data !== null) {
+          const informationDialogRef = this.commonService.informationDiaLogService(
+            "",
+            this.dataTranslate.DANHMUC.coquanquanly.nameofobject + " (" + data.tencoquanquanly + ") " + this.dataTranslate.DANHMUC.coquanquanly.informedContentOfUnDeletedDialog,
+            this.dataTranslate.DANHMUC.coquanquanly.informedDialogTitle,
           );
+        } else {
+            await this.dmFacadeService
+            .getDmCoQuanQuanLyService()
+            .deleteItem({ idCoquanquanly: this.selectedItem.idcoquanquanly })
+            .subscribe(
+              () => this.getAllCoQuanQuanLy(),
+              (error: HttpErrorResponse) => {
+                this.commonService.showeNotiResult(error.message, 2000);
+              },
+              () =>
+                this.commonService.showeNotiResult(
+                  this.dataTranslate.COMMON.default.successDelete,
+                  2000
+                )
+            );
+        }
       }
     });
   }
